log: trace_id is 64 bit now. Exporting of trace_id to eblob added.

+++ bindings/cpp/log.cpp
@@ -21,7 +21,7 @@
 
 using namespace ioremap::elliptics;
 
-extern __thread uint32_t trace_id;
+extern __thread uint64_t trace_id;
 
 class ioremap::elliptics::logger_data {
 	public:
